Added GrowlApplicationBridge, which implements (at the moment) a single public method:
/*
+ (BOOL)launchGrowlIfInstalledNotifyingTarget:(id)target selector:(SEL)selector context:(id)context
Returns YES (TRUE) if the Growl helper app began launching.
Returns NO (FALSE) and performs no other action if the Growl prefPane is not properly installed.
GrowlApplicationBridge will send "selector" to "target" when Growl is ready for use (this will only occur when it also returns YES).
	Note: selector should take a single argument; this is to allow applications to have context-relevent information passed back. It is perfectly
	acceptable for context to be NULL.
*/

Made a few minor BeepController fixes while implementing the use of this method.  When BeepController's "Register" checkbox is checked, Growl will now launch if it is not already launched (assuming the prefPane is properly installed).
Note that you need to rebuild the prefpane for this to work, too, as a few changes were made to the GrowlHelperApplication in this implementation.

--HG--
extra : convert_revision : svn%3A99687598-2e92-11dd-8019-fe7b9d601d1b/trunk%4089