Greatly simplified the behavior of the GrowlAppBridge.  (This may or may not be easy to do with the Carbon bridge.. it certainly continues to function the way it is, and I'll leave that to someone who knows Carbon to decide). The GrowlAppBridge is now initialized through a single method: - (void)  setGrowlDelegate:(NSObject<GrowlAppBridgeDelegate> *)inDelegate
The growl delegate is required to conform to the GrowlAppBridgeDelegate protocol, which means at present that it must implement
- (NSString *) growlAppName
and
- (NSDictionary *) growlRegistrationDict
to return the information used for registering with Growl and, in the case of the app name, in posting notifications for display.

It may also implement methods from the GrowlAppBridgeDelegate_InformalProtocol to return alternate growl app icon data, be told when growl is ready, be informed when a growl notification is clicked, and so on.
The methods have sparse/no documentation, but I'll fix that.

Renamed the *ApplicationBridge* files to *AppBridge* to match the class naming.  Updated the Growl.xcode project to reflect this, and created Growl.framework, which apps will include and link against to make easy use of the GrowlApplicationBridge.
There is no need for an installed framework which apps can weak link against... Growl.framework is 64 kb compiled and uncompressed, and it's much simpler for application writers to make use of it directly than to weak link against it; additionally, this means we don't need a installer but can instead distribute just the pref pane.  Discussed this with Tick and he gave it the go-ahead.

--HG--
extra : convert_revision : svn%3A99687598-2e92-11dd-8019-fe7b9d601d1b/trunk%401069