Growl-WithInstaller.framework now lives up to its name :)
The first time an application calls GrowlAppBridge's notifyWithTitle method, if Growl did not launch previously because it was not installed, the user is prompted to install it.
The installation prompt window can be customized by the growlDelegate, which (if it implements the methods) will be prompted for the window title and for the text to be displayed.  If the app does not provide one, Growl provides default text, which will be localized.
The buttons on the window are also localized, purely within the code, so no nib localization will be necessary (keeping the Growl.framework size to a minimum).
Growl.prefPane is stored within the Growl-WithInstaller.framework in archived form (zipped using /usr/bin/zip, decompressed using /usr/bin/unzip if needed).

Growl-WithInstaller also supports automatic notification and subsequent upgrading of Growl.  If the version packaged with the framework is newer than the currently installed version, the user will be notified; as with installation, the implementing application can choose to provide customized text, and if it doesn't Growl will provide defaults.

This commit also provides miscellaneous improvements to the GrowlAppBridge regarding notifications posted while Growl is initially launching.

The project changes, in addition to adding the needed files, build GrowlHelperApp with -force_flat_namespace, which I found fixed undefined symbol problems when building in deployment on one of my systems.

--HG--
extra : convert_revision : svn%3A99687598-2e92-11dd-8019-fe7b9d601d1b/trunk%401092