Partially semi-reverted r3553, which broke GAB's attempts to get some kind of app name.
r3553 makes GAB only try alternate sources of an app name if there's a delegate that implements the method and it returns nil. This doesn't make a lot of sense; the main reason to try alternate sources is when the delegate *doesn't* exist or implement the method; if it does exist and implement the method, it almost certainly will not return nil.

Now, as before r3553, GAB will try the other sources of app name even when there's no delegate or the delegate doesn't implement the method.

I think I know where Ingmar was going with it: Don't fallback on source #3 when source #1 succeeds. I've gone ahead and kept that; what I changed was that it was not falling back on source ''#2'' when source #1 failed.

--HG--
extra : convert_revision : svn%3A99687598-2e92-11dd-8019-fe7b9d601d1b/trunk%404281