Delegate the opening of .growlRegDict files to a new GrowlPropertyListFilePathway, and only terminate the app one second after the last file is processed.
This fixes, GrowlHelperApp-side, the bug where a bunch of .growlRegDict files get left in the Temporary Items folder. This was caused by GAB launching the .growlRegDict files with the GHA in ~/Library, when the GHA in /Library was the one running, and by GHA-in-~/Library bailing before it processed the files. A GHA with this fix will (a) process all files it was launched with, not just the first one, and (b) delete the files after processing them, as it's supposed to.

--HG--
extra : convert_revision : svn%3A99687598-2e92-11dd-8019-fe7b9d601d1b/trunk%404803