Set the maxWidth variable when we find a larger image rep than the one we had previously. Otherwise, we simply end up with the last image rep in the array; the width comparisons have no effect.
This bug would cause the WebKit display to sometimes show a '?' when the source image is in the IconFamily format, because IconFamily images can contain a 1-bit representation, and that representation is usually last when it's present. This fix is only partial: it's still possible for this method to let a 1-bit rep win, which will cause the '?' again.

--HG--
extra : convert_revision : svn%3A99687598-2e92-11dd-8019-fe7b9d601d1b/trunk%404820