Fixed what was likely one of the major causes of Growl-to-Growl local network forwarding failing.  Previously, when forwarding was configured, the destination address and port were stored as a preference and used forever-after.  That's wrong when destinations aren't on static IPs, which is by far the most common situation.
We now do a proper resolve to get the IP and port of the advertised Growl TCP service.

(Note: This doesn't update the preference code to not store this information;  the address information just won't be used. The network preferences pane could use a lot of work under the hood, as it's clearly not complete.)

--HG--
extra : convert_revision : svn%3A99687598-2e92-11dd-8019-fe7b9d601d1b/trunk%404912